- Dodanie edycji oraz osblugi demand & capacity - Autoscrollowanie logów

+++ implementacja/serwer/src/main/groovy/pl/mgr/vrp/model/VRPCustomer.groovy
@@ -3,6 +3,6 @@ package pl.mgr.vrp.model
 class VRPCustomer extends VRPLocation {
 
     Map restrictions
-    Integer demand
+    double demand
 
 }
